small refactor three-way-switch component
