+++ src/components/three-way-switch/three-way-switch.component.tsx
@@ -1,28 +1,7 @@
 import React from 'react';
 import { ThreeWayToggleType } from '../../view/to-do/to-do.view';
-import { Switch, SwitchLabel, SwitchRadio, SwitchSelection } from './three-way-switch.styles';
-
-const titleCase = (str: string) =>
-    str.split(/\s+/).map(w => w[0].toUpperCase() + w.slice(1)).join(' ');
-
-type ClickableLabelType = {
-    title: ThreeWayToggleType;
-    onChange: (title: ThreeWayToggleType) => void;
-    id: number
-}
-
-const ClickableLabel = ({ title, onChange, id }: ClickableLabelType) =>
-    <SwitchLabel onClick={() => onChange(title)} className={"" + id}>
-        {titleCase(title)}
-    </SwitchLabel>;
-
-type ConcealedRadioType = {
-    value: ThreeWayToggleType;
-    selected: ThreeWayToggleType;
-}
-
-const ConcealedRadio = ({ value, selected }: ConcealedRadioType) =>
-    <SwitchRadio type="radio" name="switch" checked={selected === value} readOnly />;
+import { ClickableLabel, ConcealedRadio } from './switch.utils-components';
+import { Switch, SwitchSelection } from './three-way-switch.styles';
 
 type ThreeWaySwitchType = {
     values: ThreeWayToggleType[];
